test(api-diff): consolidate class optionality coverage
