+++ .changeset/retain-class-optionality.md
@@ -1,0 +1,5 @@
+---
+"@effect/api-diff": patch
+---
+
+Retain optional class members in API snapshots.
